modules/video/plex/requirements.txt
plexapi>=3.1.0
paho-mqtt>=1.4.0
websocket-client
pyyaml